backend/requirements.txt
streamlit
pyserial